+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 - Build statically linked executables.
 - Rename "service" muss concept to "module" to improve clarity.
   Show deprecation warnings when old names are used.
+- Don't error for any failure to precreate mount points
+  (give docker(-compose) the chance to deal with it).
 
 # v0.7 - 2020-02-28